requirements.txt
absl-py==0.9.0
alembic==1.4.2
altgraph==0.17
apache-airflow==1.10.15
apispec==1.3.3
appdirs==1.4.4
argcomplete==1.11.1
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
auto-py-to-exe==2.6.6
Babel==2.8.0
backcall==0.1.0
bleach==3.3.0
bottle==0.12.19
bottle-websocket==0.2.9
cached-property==1.5.1
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.31
# cryptography==2.1.4
# cv==1.0.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
Django==3.0.14
docutils==0.16
Eel==0.11.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.1
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.16.0
gast==0.3.3
gevent==1.4.0
gevent-websocket==0.10.1
google-pasta==0.2.0
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.29.0
gunicorn==19.10.0
h5py==2.10.0
idna==2.6
image==1.5.32
importlib-metadata==1.6.0
imutils==0.5.3
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.3
joblib==0.14.1
json-merge-patch==0.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.2
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.22.3
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.4.5
notebook==6.1.5
numpy==1.18.2
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.2
pendulum==1.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.1
prison==0.1.3
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
# pycairo==1.16.2
pycrypto==2.6.1
Pygments==2.7.4
# pygobject==3.26.1
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
pytesseract==0.3.3
# python-apt===1.6.5-ubuntu0.3-elementary16-ubuntu5.1.5.1
python-daemon==2.1.2
python-dateutil==2.8.1
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
pyxdg==0.26
PyYAML==5.4
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
scour==0.36
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.15
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.3
sqlparse==0.3.1
tabulate==0.8.7
tb-nightly==1.15.0a20190806
tenacity==4.12.0
tensorflow==2.4.0
termcolor==1.1.0
terminado==0.8.3
# tesseract-ocr==0.0.1
testpath==0.4.4
text-unidecode==1.2
tf-estimator-nightly==1.14.0.dev2019080601
thrift==0.13.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.8
virtualenv==20.0.23
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
whichcraft==0.6.1
widgetsnbextension==3.5.1
wordcloud==1.6.0
wrapt==1.11.2
WTForms==2.2.1
zipp==3.1.0
zope.deprecation==4.4.0